test-requirements.txt
pytest
coverage
flake8
pytest-mock
awsiotsdk